Remove V9GeometryFilterer class
